Calling starkcheck in send page
