+++ .env.development
@@ -5,7 +5,7 @@ NEXT_PUBLIC_STARKCHECK_API_ENDPOINT=http://127.0.0.1:6060
 
 # The variables below are linked to already generated contracts on testnet.
 # Update them only if you want to use your own deployer/proxy/contract implementation
-CONTRACT_ACCOUNT_CLS_HASH="0x06928805c504ca234006781feb068b0f948660861ce5cf0f3c159c43377d7ca5"
+CONTRACT_ACCOUNT_CLS_HASH="0x0513bd1acc5476c75ba04fc1a30263e913b0f9628d31662e0cfc1f9e8f90b2a8"
 PROXY_CLS_HASH="0x04be79b3904b4e2775fd706fa037610b41d8f8708ce298aac3a470badf68176d"
 STARKSIGNER_CLS_HASH="0x2b59e7b163ac41885672f679a156117c5347628273ffc9ca2bcd77a024a176c"
 WEBAUTHN_CLS_HASH="0x0769efb07a6e3508fe229895fe5a0033dc59de4be3e26e7a2bfe9c1d48ec90e2"
